Django_app/myproject/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
cssbeautifier==1.15.4
Django==5.2.3
django-allauth==65.9.0
django-redis==5.4.0
djlint==1.36.4
EditorConfig==0.17.1
idna==3.10
jsbeautifier==1.15.4
json5==0.12.0
pathspec==0.12.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.0
PyYAML==6.0.2
redis==6.2.0
regex==2024.11.6
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.14.0
urllib3==2.4.0
